requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
appdirs==1.4.4
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.5
dill==0.3.7
docker==7.0.0
entrypoints==0.4
etils==1.8.0
fastobo==0.12.3
filelock==3.9.0
Flask==3.0.2
flax==0.8.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.42
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==21.2.0
huggingface-hub==0.17.3
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lark==1.1.9
llvmlite==0.40.1
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.11.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.15
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.57.1
numpy==1.24.4
opt-einsum==3.3.0
optax==0.2.1
orbax-checkpoint==0.5.7
packaging==23.2
pandas==2.0.3
pillow==10.2.0
polars==0.19.7
protobuf==3.19.6
pyarrow==15.0.0
Pygments==2.17.2
pyparsing==3.1.2
pyteomics==4.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
querystring-parser==1.2.4
requests==2.31.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
spectrum-utils==0.4.2
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
tensorstore==0.1.56
threadpoolctl==3.4.0
toolz==0.12.1
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tqdm==4.65.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
